fix: refactoring folder structure for semester
