+++ .nvmrc
@@ -1,1 +1,1 @@
-v22.12.0
+20.9.0
