+++ packages/api/src/feature/semester/semester.module.ts
@@ -1,7 +1,7 @@
 import { Module } from "@nestjs/common";
 import { DrizzleModule } from "@sparcs-students/api/drizzle/drizzle.module";
-import { SemesterRepository } from "./semester.repository";
-import { SemesterPublicService } from "./semester.public.service";
+import { SemesterRepository } from "./repository/semester.repository";
+import { SemesterPublicService } from "./service/semester.public.service";
 
 @Module({
   providers: [SemesterRepository, SemesterPublicService],
